// src/enums/role-id.ts
export enum RoleId {
  Unknown = 'Unknown',
  Merlin = 'Merlin',
  Percival = 'Percival',
  Servant_1 = 'Servant_1',
  Servant_2 = 'Servant_2',
  Servant_3 = 'Servant_3',
  Servant_4 = 'Servant_4',
  Servant_5 = 'Servant_5',
  Mordred = 'Mordred',
  Morgana = 'Morgana',
  Assassin = 'Assassin',
  Oberon = 'Oberon',
  Minion_1 = 'Minion_1',
  Minion_2 = 'Minion_2',
  Minion_3 = 'Minion_3',
}